backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
python-decouple==3.8
psycopg2-binary==2.9.10
drf-yasg==1.21.10
celery==5.5.3
django-filter==24.1
redis==5.0.1 